requirements.txt
absl-py==0.3.0
APScheduler==3.5.1
astor==0.7.1
attrs==18.1.0
Automat==0.7.0
bleach==1.5.0
boto3==1.5.20
botocore==1.8.50
certifi==2018.4.16
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
decorator==4.3.0
dill==0.2.8.2
docopt==0.6.2
docutils==0.14
en-core-web-md==2.0.0
en-core-web-sm==2.0.0
fakeredis==0.10.3
fbmessenger==5.3.0
Flask==1.0.2
Flask-Cors==3.0.6
future==0.16.0
fuzzywuzzy==0.16.0
gast==0.2.0
gevent==1.2.2
graphviz==0.8.4
greenlet==0.4.14
grpcio==1.13.0
h5py==2.8.0
html5lib==0.9999999
humanfriendly==4.16.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonpickle==0.9.6
jsonschema==2.6.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
klein==17.10.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0
mattermostwrapper==2.1
msgpack==0.5.6
msgpack-numpy==0.4.3.1
msgpack-python==0.5.4
murmurhash==0.28.0
networkx==2.1
numpy==1.14.4
packaging==17.1
pathlib==1.0.1
plac==0.9.6
preshed==1.0.1
protobuf==3.6.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pykwalify==1.6.0
pyparsing==2.2.0
PySocks==1.6.8
python-crfsuite==0.9.5
python-dateutil==2.7.3
python-Levenshtein==0.12.0
python-telegram-bot==10.1.0
pytz==2018.5
PyYAML==3.12
rasa-core==0.10.2
rasa-nlu==0.12.3
redis==2.10.6
regex==2017.4.5
requests==2.18.4
ruamel.yaml==0.15.44
s3transfer==0.1.13
scikit-learn==0.19.2
scipy==1.1.0
simplejson==3.13.2
six==1.11.0
sklearn==0.0
sklearn-crfsuite==0.3.6
slackclient==1.2.1
spacy==2.0.12
tabulate==0.8.2
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
thinc==6.10.3
toolz==0.9.0
tqdm==4.19.5
twilio==6.15.1
Twisted==18.7.0
typing==3.6.2
tzlocal==1.5.1
ujson==1.35
urllib3==1.22
websocket-client==0.48.0
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.5.0
